requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
dj-database-url==0.5.0
Django==3.1.1
django-bootstrap4==2.2.0
django-countries==6.1.3
django-filter==2.3.0
django-heroku==0.3.1
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.2.0